Renamed and deleted incorrect packages and deleted not needed classes. Classes were transferred to the correct packages.
