+++ src/main/java/com/example/Library/config/JwtConfigurer.java
@@ -1,6 +1,6 @@
-package com.example.Library.security;
+package com.example.Library.config;
 
-import org.springframework.beans.factory.annotation.Autowired;
+import com.example.Library.security.JwtTokenFilter;
 import org.springframework.security.config.annotation.SecurityConfigurerAdapter;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.web.DefaultSecurityFilterChain;
