+++ src/main/java/com/example/Library/controller/AdminRestController.java
@@ -1,11 +1,10 @@
-package com.example.Library.rest;
+package com.example.Library.controller;
 
 import com.example.Library.exceptions.*;
 import com.example.Library.model.Book;
-import com.example.Library.model.enums.BookStatus;
-import com.example.Library.model.enums.Language;
+import com.example.Library.enums.BookStatus;
+import com.example.Library.model.Language;
 import com.example.Library.repositories.BookRepository;
-import com.example.Library.services.BookService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
@@ -16,9 +15,6 @@ import java.util.Optional;
 @RestController
 @RequestMapping("api/v1/Admins")
 public class AdminRestController {
-
-
-
 
     private final BookRepository bookRepository;
