--- src/main/java/com/example/Library/controller/AuthController.java
@@ -1,22 +1,0 @@
-package com.example.Library.controller;
-
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-@RequestMapping("/auth")
-public class AuthController {
-
-    @GetMapping("/login")
-    public String getLoginPage(){
-        return "login";
-    }
-
-
-    @GetMapping("/success")
-    public String getSuccessPage(){
-        return "success";
-    }
-}
