+++ src/main/java/com/example/Library/dto/AuthenticationRequestDTO.java
@@ -1,4 +1,4 @@
-package com.example.Library.rest;
+package com.example.Library.dto;
 
 import lombok.Data;
