+++ src/main/java/com/example/Library/model/Book.java
@@ -1,7 +1,6 @@
 package com.example.Library.model;
 
-import com.example.Library.model.enums.BookStatus;
-import com.example.Library.model.enums.Language;
+import com.example.Library.enums.BookStatus;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
