+++ src/main/java/com/example/Library/security/JwtTokenFilter.java
@@ -1,5 +1,6 @@
 package com.example.Library.security;
 
+import com.example.Library.exceptions.JwtAuthenticationException;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Component;
