+++ src/main/java/com/example/Library/services/UserDetailsServiceImpl.java
@@ -1,7 +1,8 @@
-package com.example.Library.security;
+package com.example.Library.services;
 
 import com.example.Library.model.User;
 import com.example.Library.repositories.UserRepository;
+import com.example.Library.security.SecurityUser;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
