--- src/main/resources/templates/success.html
@@ -1,6 +1,0 @@
-<div class="container">
-    <h1>Congrats!!! You did it!</h1>
-    <form action="/auth/logout" method="POST">
-        <button type="submit">Logout</button>
-    </form>
-</div>
